requirements.txt
python-dotenv
Django
django-tables2
django-crispy-forms
# django-silk
tablib
boto3
crispy-tailwind
